Added Easy Spin platform but didn't compile....
NPM and the entire ecosystem is an unstable house of cards. Despite
zero code or depency changes, somebody upstream decided to make breaking
changes to one or more dependencies, or build tools.

I can no longer get this to build on Mac. I do not have time to
rewrite or modify this.

TODO:
* `npm run dev`
* Verify:
  * "Easy Spin" platform is listed at the top of the index page. Link works.
  * "Distortion algorithms" appears under distortion cateogory. Download link works
* `npm run generate`
* commit new build
* submit pull request.